+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### master
 - if the pane is too narrow do not show sidebar
+- improve sidebar max size handling
 
 ### v0.2.0, Sep 03, 2014
 - automatic sidebar opening for another sidebar type